+++ IBAnimatable/AnimatorFactory.swift
@@ -26,7 +26,11 @@ struct AnimatorFactory {
       return SystemFlipAnimator(fromDirection: direction, transitionDuration: transitionDuration)
     case .SystemPageCurl(let direction):
       return SystemPageCurlAnimator(fromDirection: direction, transitionDuration: transitionDuration)
-    }
+    case .SystemSuckEffect:
+      return SystemSuckEffectAnimator(transitionDuration: transitionDuration)
+    case .SystemCameraIris(let hollowState):
+      return SystemCameraIrisAnimator(hollowState: hollowState, transitionDuration: transitionDuration)
+  }
   }
   
 }