+++ IBAnimatable/TransitionHollowState.swift
@@ -1,0 +1,17 @@
+//
+//  TransitionCameraIrisHollowState.swift
+//  IBAnimatableApp
+//
+//  Created by Tom Baranes on 30/03/16.
+//  Copyright © 2016 Jake Lin. All rights reserved.
+//
+
+import Foundation
+
+/**
+ Transition Hollow State
+ */
+public enum TransitionHollowState {
+  case Open
+  case Close
+}