+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 #### Enhancements
 
 - `AnimatableLabel` now conforms `BorderDesignable`
+- Add `SystemSuckEffect` transition animations [#126](https://github.com/JakeLin/IBAnimatable/issues/126)
+- Add `SystemCameraIrisAnimator` to support `SystemCameraIrisAnimator(Open)` and `SystemCameraIrisAnimator(Close)` transition animations [#126](https://github.com/JakeLin/IBAnimatable/issues/126)
 
 #### Bugfixes